% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGRanges.R
\name{CGRanges}
\alias{CGRanges}
\title{CGRanges S7 Class
An S7 class representing a CGRanges object.}
\usage{
CGRanges(.ptr = .Call(RC_cr_init), indexed = FALSE)
}
\description{
CGRanges S7 Class
An S7 class representing a CGRanges object.
}
