% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGRanges.R
\name{CGRangesOverlap}
\alias{CGRangesOverlap}
\title{CGRangesOverlap S7 Class
Inherits from CGRanges, adds 'hits' field for overlap indices}
\usage{
CGRangesOverlap(.ptr = .Call(RC_cr_init), indexed = FALSE, hits = list())
}
\description{
CGRangesOverlap S7 Class
Inherits from CGRanges, adds 'hits' field for overlap indices
}
